geotiff tests: consolidate georef and rotation cluster (#2435)
Fold twelve top-level georef / no-georef / rotation test files into
xrspatial/geotiff/tests/read/test_georef.py. Read-side and write-side
cases share the rotated-affine and no-georef-marker fixtures, so the
write-side tests stay in the same module under a write-side heading
rather than splitting into write/test_georef.py.

GPU gating moves to the shared requires_gpu marker and gpu_available
probe from _helpers/markers.py. Per-issue rotated-TIFF and VRT writers
keep their distinct byte layouts and are suffixed by issue number to
avoid collision. The status suite imports _write_rotated_tiff from
read/test_crs.py, which is left untouched.

CLUSTER_AUDIT_GEOREF.md maps every old file to its new section; it is
deleted in a pre-merge commit per epic #2424.
